refactor: :recycle: update path files for the import of `ContactsList` to `index` and `profile` in `pages`
