+++ src/pages/index.jsx
@@ -1,16 +1,16 @@
 import { Suspense } from "react";
 import { Await, useLoaderData } from "react-router-dom";
-import ContactsList from "../components/contacts-list";
+import ContactsList from "../components/contacts/index.jsx";
 import Loading from "../components/loading";
 
-export default function Index(){
-    const contacts = useLoaderData();
+export default function Index() {
+  const contacts = useLoaderData();
 
-    return (
-        <Suspense fallsback={<Loading />}> 
-        <Await resolve={contacts}>
-            <ContactsList />
-            </Await>
-            </Suspense>
-    );
+  return (
+    <Suspense fallsback={<Loading />}>
+      <Await resolve={contacts}>
+        <ContactsList />
+      </Await>
+    </Suspense>
+  );
 }
